requirements.txt
Django==1.7.1
South==1.0.1
coverage==3.7.1
django-admin-tools==0.5.2
django-bootstrap3==4.11.0
django-debug-toolbar==1.2.2
django-extensions==1.4.5
gnureadline==6.3.3
ipython==2.3.0
psycopg2==2.5.4
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
